+++ src/anomalib/models/image/__init__.py
@@ -23,9 +23,11 @@ Example:
     >>> predictions = engine.predict(model=model, datamodule=datamodule)  # doctest: +SKIP
 
 Available Models:
-    - :class: `AnomalyDINO`: Boost Memorybank Models with DINOv2
+    - :class:`AnomalyVFM`: Transforming Vision Foundation Models into Zero-Shot Anomaly Detectors
+    - :class:`AnomalyDINO`: Boost Memorybank Models with DINOv2
     - :class:`Cfa`: Contrastive Feature Aggregation
     - :class:`Cflow`: Conditional Normalizing Flow
+    - :class:`CFM`: Crossmodal Feature Mapping
     - :class:`Csflow`: Conditional Split Flow
     - :class:`Dfkde`: Deep Feature Kernel Density Estimation
     - :class:`Dfm`: Deep Feature Modeling
@@ -51,8 +53,10 @@ Available Models:
 """
 
 from .anomaly_dino import AnomalyDINO
+from .anomalyvfm import AnomalyVFM
 from .cfa import Cfa
 from .cflow import Cflow
+from .cfm import CFM
 from .csflow import Csflow
 from .dfkde import Dfkde
 from .dfm import Dfm
@@ -64,6 +68,8 @@ from .fastflow import Fastflow
 from .fre import Fre
 from .ganomaly import Ganomaly
 from .general_ad import GeneralAD
+from .glass import Glass
+from .inp_former import InpFormer
 from .l2bt import L2BT
 from .padim import Padim
 from .patchcore import Patchcore
@@ -77,9 +83,11 @@ from .vlm_ad import VlmAd
 from .winclip import WinClip
 
 __all__ = [
+    "AnomalyVFM",
     "AnomalyDINO",
     "Cfa",
     "Cflow",
+    "CFM",
     "Csflow",
     "Dfkde",
     "Dfm",
@@ -91,6 +99,8 @@ __all__ = [
     "Fre",
     "Ganomaly",
     "GeneralAD",
+    "Glass",
+    "InpFormer",
     "L2BT",
     "Padim",
     "Patchcore",